pyproject.toml
[tool.poetry]
name = "sequencehandler"
version = "0.1.0"
description = ""
authors = ["mervebduman"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
biopython = "^1.83"
pyyaml = "^6.0.1"
pytest = "^8.2.0"
levenshtein = "^0.25.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sequencehandler = "sequencehandler.main:main"